Copy the sources in addition to the image
At least in Pharo7 the sources need to be accessible for it to work. If the sources are located next to the image (in contrast to the VM) we need to copy/link them to the temporary directory. This is necessary for runs with smalltalkCI but should match common deployments as well.

Another option was to avoid creating a temporary directory but then the clean-up is complicated to impossible (which ombus-session must go?).